refactor!: Migrate Config model to TS
